+++ src/Calamity/Types/Model/Guild/Emoji.hs
@@ -2,7 +2,8 @@
 module Calamity.Types.Model.Guild.Emoji
     ( Emoji(..)
     , Partial(PartialEmoji)
-    , RawEmoji(..) ) where
+    , RawEmoji(..)
+    , emojiAsRawEmoji ) where
 
 import           Calamity.Internal.AesonThings
 import           Calamity.Internal.Utils         ()
@@ -11,6 +12,7 @@ import           Calamity.Types.Model.User
 import           Calamity.Types.Snowflake
 
 import           Data.Aeson
+import qualified Data.Text.Lazy                  as L
 import           Data.Text.Lazy                  ( Text )
 import           Data.Vector.Unboxing            ( Vector )
 
@@ -18,6 +20,7 @@ import           GHC.Generics
 
 import           TextShow
 import qualified TextShow.Generic                as TSG
+import Data.Generics.Product
 
 data Emoji = Emoji
   { id            :: Snowflake Emoji
@@ -34,20 +37,42 @@ data Emoji = Emoji
   deriving ( FromJSON ) via WithSpecialCases '["user" `ExtractFieldFrom` "id"] Emoji
   deriving ( HasID Emoji ) via HasIDField "id" Emoji
 
+emojiAsRawEmoji :: Emoji -> RawEmoji
+emojiAsRawEmoji = CustomEmoji . upcast
+
 data instance Partial Emoji = PartialEmoji
   { id   :: Snowflake Emoji
   , name :: Text
+  , animated :: Bool
   }
-  deriving ( Eq, Show, Generic )
-  deriving ( TextShow ) via TSG.FromGeneric (Partial Emoji)
-  deriving ( ToJSON, FromJSON ) via CalamityJSON (Partial Emoji)
+  deriving ( Eq, Generic )
+  deriving ( ToJSON ) via CalamityJSON (Partial Emoji)
+  deriving ( FromJSON ) via WithSpecialCases
+      '["animated" `IfNoneThen` DefaultToFalse] (Partial Emoji)
   deriving ( HasID Emoji ) via HasIDField "id" (Partial Emoji)
+
+instance Show (Partial Emoji) where
+  show PartialEmoji { id, name, animated } =
+    "<" <> a <> ":" <> L.unpack name <> ":" <> show id <> ">"
+    where a = if animated then "a" else ""
+
+instance TextShow (Partial Emoji) where
+  showb PartialEmoji { id, name, animated } =
+    "<" <> a <> ":" <> fromLazyText name <> ":" <> showb id <> ">"
+    where a = if animated then "a" else ""
 
 data RawEmoji
   = UnicodeEmoji Text
   | CustomEmoji (Partial Emoji)
-  deriving ( Eq, Show, Generic )
-  deriving ( TextShow ) via TSG.FromGeneric RawEmoji
+  deriving ( Eq, Generic )
+
+instance Show RawEmoji where
+  show (UnicodeEmoji v) = L.unpack v
+  show (CustomEmoji p) = show p
+
+instance TextShow RawEmoji where
+  showb (UnicodeEmoji v) = fromLazyText v
+  showb (CustomEmoji p) = showb p
 
 instance ToJSON RawEmoji where
   toJSON (CustomEmoji e) = object ["emoji" .= e]
@@ -56,8 +81,9 @@ instance ToJSON RawEmoji where
 instance FromJSON RawEmoji where
   parseJSON = withObject "RawEmoji" $ \v -> do
     m_id :: Maybe (Snowflake Emoji) <- v .:? "id"
+    anim <- v .:? "animated" .!= False
     name :: Text <- v .: "name"
 
     pure $ case m_id of
-      Just id -> CustomEmoji $ PartialEmoji id name
+      Just id -> CustomEmoji $ PartialEmoji id name anim
       Nothing -> UnicodeEmoji name